Fix logic for nested interpolation Co-authored-by: Steven Murawski <steven.murawski@microsoft.com>
