+++ src/tests/service/test_service_environment.py
@@ -1,5 +1,4 @@
-from unittest import TestCase, mock
-import os
+from unittest import TestCase
 
 from src.pycomposefile import compose_file
 
@@ -20,13 +19,3 @@ class TestComposeServiceEnvironment(TestCase):
         self.assertEqual(compose_file.services["frontend"].environment["RACK_ENV"], "development")
         self.assertEqual(compose_file.services["frontend"].environment["SHOW"], "true")
         self.assertIsNone(compose_file.services["frontend"].environment["USER_INPUT"])
-
-    @mock.patch.dict(os.environ, {"RACK_ENV": "test"})
-    @mock.patch.dict(os.environ, {"test": "https://127.0.0.1"})
-    @mock.patch.dict(os.environ, {"VERSION": "release"})
-    def test_service_with_os_environment_vars(self):
-        compose_file = ComposeGenerator.get_compose_with_os_environment_vars()
-
-        self.assertEqual(compose_file.services["frontend"].environment["RACK_ENV"], "test")
-        self.assertEqual(compose_file.services["frontend"].environment["URL"], "https://127.0.0.1")
-        self.assertEqual(compose_file.services["frontend"].environment["VERSION"], "release")
